Put q&a component in about.tsx
